rays: Move ColorSpace GL function to opengl/color_space
Extract ColorSpace_get_gl_format_and_type() from src/color_space into
src/opengl/color_space, removing GL dependency from the common code.
