+++ rays/src/ios/bitmap.mm
@@ -6,7 +6,7 @@
 #import <MobileCoreServices/UTCoreTypes.h>
 #include <xot/util.h>
 #include "rays/exception.h"
-#include "../color_space.h"
+#include "../opengl/color_space.h"
 #include "../font.h"
 #include "../texture.h"
 #include "../frame_buffer.h"
