+++ rays/src/opengl/color_space.h
@@ -1,11 +1,11 @@
 // -*- c++ -*-
 #pragma once
-#ifndef __RAYS_SRC_COLOR_SPACE_H__
-#define __RAYS_SRC_COLOR_SPACE_H__
+#ifndef __RAYS_SRC_OPENGL_COLOR_SPACE_H__
+#define __RAYS_SRC_OPENGL_COLOR_SPACE_H__
 
 
 #include "rays/color_space.h"
-#include "opengl/opengl.h"
+#include "opengl.h"
 
 
 namespace Rays
