+++ rays/src/osx/bitmap.mm
@@ -5,7 +5,7 @@
 #import <Cocoa/Cocoa.h>
 #include <xot/util.h>
 #include "rays/exception.h"
-#include "../color_space.h"
+#include "../opengl/color_space.h"
 #include "../font.h"
 #include "../texture.h"
 #include "../frame_buffer.h"
