+++ rays/src/win32/bitmap.cpp
@@ -7,7 +7,7 @@
 #include <stb_image_write.h>
 
 #include "rays/exception.h"
-#include "../color_space.h"
+#include "../opengl/color_space.h"
 #include "../font.h"
 #include "../texture.h"
 #include "../frame_buffer.h"
